Bake in MixinExtras artifact in UserDev Config v2
